+++ Sources/APRelay/Entrypoint.swift
@@ -8,8 +8,13 @@ struct Entrypoint {
         try LoggingSystem.bootstrap(from: &env)
 
         let app = try await Application.make(env)
-        try await configure(app)
-        try await app.execute()
-        try await app.asyncShutdown()
+        do {
+            try await configure(app)
+            try await app.execute()
+            try await app.asyncShutdown()
+        } catch {
+            try? await app.asyncShutdown()
+            throw error
+        }
     }
 }